+++ appdaemon/assets/javascript/dashboard.js
@@ -6,7 +6,7 @@ function ha_status(stream, dash, widgets)
     webSocket.onopen = function (event) 
     {
         webSocket.send(dash);
-    }
+    };
 
     webSocket.onmessage = function (event) 
     {
@@ -17,12 +17,12 @@ function ha_status(stream, dash, widgets)
                 widgets[key].on_ha_data(JSON.parse(event.data));
             }
         })
-    }
+    };
     webSocket.onclose = function (event)
     {
         //window.alert("Server closed connection")
        // window.location.reload(false); 
-    }
+    };
 
     webSocket.onerror = function (event)
     {
@@ -58,15 +58,15 @@ var inheritsFrom = function (child, parent) {
 
 var WidgetBase = function(widget_id, url, skin, parameters, monitored_entities, callbacks)
 {
-    child = this
-    child.url = url
+    child = this;
+    child.url = url;
     
     // Function definitions
     
     this.set_field = function(self, field, value)
     {
         self.ViewModel[field](value)
-    }
+    };
     
     this.format_number = function(self, value)
     {
@@ -95,7 +95,7 @@ var WidgetBase = function(widget_id, url, skin, parameters, monitored_entities, 
             value = value.toString().replace(".", ",")
         }
         return value
-    }
+    };
 
     
     this.map_state = function(self, value)
@@ -116,12 +116,12 @@ var WidgetBase = function(widget_id, url, skin, parameters, monitored_entities, 
             state = value
         }
         return (state)
-    }
+    };
     
     this.set_icon = function(self, field, value)
     {
         self.ViewModel[field](value.split("-")[0] + ' ' + value)
-    }
+    };
     
     this.get_state = function(child, base_url, entity)
     {
@@ -157,7 +157,7 @@ var WidgetBase = function(widget_id, url, skin, parameters, monitored_entities, 
                             {
                                 child.entity_state = {}
                             }
-                            child.entity_state[entity.entity] = new_state
+                            child.entity_state[entity.entity] = new_state;
                             entity.initial(child, new_state)
                         }
                     },
@@ -167,7 +167,7 @@ var WidgetBase = function(widget_id, url, skin, parameters, monitored_entities, 
                     }
                   
         });
-    }
+    };
    
     this.on_ha_data = function(data)
     {
@@ -179,18 +179,18 @@ var WidgetBase = function(widget_id, url, skin, parameters, monitored_entities, 
             {
                 if (monitored_entities[i].entity == entity)
                 {
-                    this.entity_state[entity] = data.data.new_state
+                    this.entity_state[entity] = data.data.new_state;
                     monitored_entities[i].update(this, data.data.new_state)
                 }
             }
         }
-    }
+    };
     
     this.call_service = function(child, args)
     {
         service_url = child.url + "/" + "call_service";
-        $.post(service_url, args); 
-    }
+        $.post(service_url, args);
+    };
 
     // Initialization
     
@@ -218,17 +218,17 @@ var WidgetBase = function(widget_id, url, skin, parameters, monitored_entities, 
     
     // Create and initialize bindings
     
-    child.ViewModel = {}
+    child.ViewModel = {};
     
     Object.keys(parameters.fields).forEach(function(key,index)
     {
         child.ViewModel[key] = ko.observable()
     });
 
-    child.css = {}
+    child.css = {};
     Object.keys(parameters.css).forEach(function(key,index) 
     {
-        child.css[key] = parameters.css[key]
+        child.css[key] = parameters.css[key];
         child.ViewModel[key] = ko.observable()
     });
     
@@ -237,10 +237,10 @@ var WidgetBase = function(widget_id, url, skin, parameters, monitored_entities, 
         child.ViewModel[key] = ko.observable()
     });
     
-    child.icons = {}
+    child.icons = {};
     Object.keys(parameters.icons).forEach(function(key,index) 
     {
-        child.icons[key] = parameters.icons[key]
+        child.icons[key] = parameters.icons[key];
         child.ViewModel[key] = ko.observable()
     });
     
@@ -267,4 +267,4 @@ var WidgetBase = function(widget_id, url, skin, parameters, monitored_entities, 
     {
         child.ViewModel[key](parameters.static_icons[key].split("-")[0] + ' ' + parameters.static_icons[key])
     });
-}
+};